feat MainMapView, FirstTimeOpenView : rewrite of main MapView and overall MainView
