+++ Nano2/CoreData/Persistence.swift
@@ -15,7 +15,7 @@ struct PersistenceController {
         let result = PersistenceController(inMemory: true)
         let viewContext = result.container.viewContext
         for _ in 0..<10 {
-            let newItem = Item(context: viewContext)
+            let newItem = LocationItem(context: viewContext)
             newItem.timestamp = Date()
         }
         do {
