+++ Nano2/Shared/CustomFont.swift
@@ -12,4 +12,5 @@ public final class CustomFont {
     static let lightText17 = Font.custom("SF-Pro-Text-Light", size: 17)
     static let roundedRegular16 = Font.custom("SF-Pro-Rounded-Regular", size: 16)
     static let expandedBold = Font.custom("SFPro-ExpandedBold", size: 28)
+    static let medium16 = Font.custom("Fredoka-Medium", size: 16)
 }
