+++ Nano2/Shared/CustomView/CustomButton.swift
@@ -7,11 +7,6 @@
 
 import SwiftUI
 
-enum CustomButton {
-    case CustomFillButton
-    case CustomBorderButton
-}
-
 extension View {
     
     func withFillButtonStyle() -> some View {
@@ -36,7 +31,7 @@ struct FillButton: ButtonStyle {
                 RoundedRectangle(cornerRadius: NumberConstant.defaultCornerRadiusSize)
                     .stroke(configuration.isPressed ? Color.gray : Color.black)
             )
-            .cornerRadius(6)
+            .cornerRadius(NumberConstant.defaultCornerRadiusSize)
             .foregroundColor(configuration.isPressed ? Color.white : Color.white)
     }
     
@@ -46,15 +41,15 @@ struct BorderButton: ButtonStyle {
     
     func makeBody(configuration: Configuration) -> some View {
         configuration.label
-            .font(Font.custom("Fredoka-Medium", size: 16))
+            .frame(maxWidth: .infinity)
+            .font(CustomFont.medium16)
             .padding()
-            .frame(height: 44)
             .background(configuration.isPressed ? Color.gray : Color.white)
             .overlay(
-                RoundedRectangle(cornerRadius: 6)
+                RoundedRectangle(cornerRadius: NumberConstant.defaultCornerRadiusSize)
                     .stroke(configuration.isPressed ? Color.gray : Color.gray)
             )
-            .cornerRadius(6)
+            .cornerRadius(NumberConstant.defaultCornerRadiusSize)
             .foregroundColor(configuration.isPressed ? Color.gray : Color.gray)
     }
