+++ Nano2/Views/MainView/MainViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  MainViewModel.swift
+//  Nano2
+//
+//  Created by Elvis Susanto on 12/20/24.
+//
+
+public final class MainViewModel {
+    
+    public init() {}
+    
+}
