+++ Nano2/Views/Untitled.swift
@@ -1,0 +1,7 @@
+//
+//  Untitled.swift
+//  Nano2
+//
+//  Created by Elvis Susanto on 12/20/24.
+//
+
